Removed ws.service.ts
Placing all websocket logic in a single file is no longer necessary because the
ng2-stompjs library handles heartbeat and retry logic for us.
