+++ ui/src/app/modules/teams/teams.module.ts
@@ -42,7 +42,6 @@ import {ArchivedBoardPageComponent} from './pages/archived-board/archived-board.
 import {SubAppComponent} from '../sub-app/sub-app.component';
 import {ActionsRadiatorViewComponent} from '../controls/actions-radiator-view/actions-radiator-view.component';
 import {InfiniteScrollModule} from 'ngx-infinite-scroll';
-import {WsService} from './services/ws.service';
 
 @NgModule({
   imports: [
@@ -68,7 +67,6 @@ import {WsService} from './services/ws.service';
     ActionItemService,
     ColumnService,
     WebsocketService,
-    WsService,
     FeedbackService,
     SaveCheckerService
   ],
